feat(v8): Remove requestData deprecations
