+++ packages/core/src/integrations/requestdata.ts
@@ -95,13 +95,7 @@ const _requestDataIntegration = ((options: RequestDataIntegrationOptions = {}) =
         return event;
       }
 
-      // The Express request handler takes a similar `include` option to that which can be passed to this integration.
-      // If passed there, we store it in `sdkProcessingMetadata`. TODO(v8): Force express and GCP people to use this
-      // integration, so that all of this passing and conversion isn't necessary
-      const addRequestDataOptions =
-        sdkProcessingMetadata.requestDataOptionsFromExpressHandler ||
-        sdkProcessingMetadata.requestDataOptionsFromGCPWrapper ||
-        convertReqDataIntegrationOptsToAddReqDataOpts(_options);
+      const addRequestDataOptions = convertReqDataIntegrationOptsToAddReqDataOpts(_options);
 
       const processedEvent = _addRequestData(event, req, addRequestDataOptions);
