+++ packages/node/src/handlers.ts
@@ -18,7 +18,6 @@ import type { Span } from '@sentry/types';
 import type { AddRequestDataToEventOptions } from '@sentry/utils';
 import {
   addRequestDataToTransaction,
-  dropUndefinedKeys,
   extractPathForTransaction,
   extractRequestData,
   isString,
@@ -29,8 +28,6 @@ import {
 
 import type { NodeClient } from './client';
 import { DEBUG_BUILD } from './debug-build';
-// TODO (v8 / XXX) Remove this import
-import type { ParseRequestOptions } from './requestDataDeprecated';
 import { isAutoSessionTrackingEnabled } from './sdk';
 
 /**
@@ -115,37 +112,9 @@ export function tracingHandler(): (
   };
 }
 
-export type RequestHandlerOptions =
-  // TODO (v8 / XXX) Remove ParseRequestOptions type and eslint override
-  // eslint-disable-next-line deprecation/deprecation
-  (ParseRequestOptions | AddRequestDataToEventOptions) & {
-    flushTimeout?: number;
-  };
-
-/**
- * Backwards compatibility shim which can be removed in v8. Forces the given options to follow the
- * `AddRequestDataToEventOptions` interface.
- *
- * TODO (v8): Get rid of this, and stop passing `requestDataOptionsFromExpressHandler` to `setSDKProcessingMetadata`.
- */
-function convertReqHandlerOptsToAddReqDataOpts(
-  reqHandlerOptions: RequestHandlerOptions = {},
-): AddRequestDataToEventOptions | undefined {
-  let addRequestDataOptions: AddRequestDataToEventOptions | undefined;
-
-  if ('include' in reqHandlerOptions) {
-    addRequestDataOptions = { include: reqHandlerOptions.include };
-  } else {
-    // eslint-disable-next-line deprecation/deprecation
-    const { ip, request, transaction, user } = reqHandlerOptions as ParseRequestOptions;
-
-    if (ip || request || transaction || user) {
-      addRequestDataOptions = { include: dropUndefinedKeys({ ip, request, transaction, user }) };
-    }
-  }
-
-  return addRequestDataOptions;
-}
+export type RequestHandlerOptions = AddRequestDataToEventOptions & {
+  flushTimeout?: number;
+};
 
 /**
  * Express compatible request handler.
@@ -154,9 +123,6 @@ function convertReqHandlerOptsToAddReqDataOpts(
 export function requestHandler(
   options?: RequestHandlerOptions,
 ): (req: http.IncomingMessage, res: http.ServerResponse, next: (error?: any) => void) => void {
-  // TODO (v8): Get rid of this
-  const requestDataOptions = convertReqHandlerOptsToAddReqDataOpts(options);
-
   const client = getClient<NodeClient>();
   // Initialise an instance of SessionFlusher on the client when `autoSessionTracking` is enabled and the
   // `requestHandler` middleware is used indicating that we are running in SessionAggregates mode
@@ -193,8 +159,6 @@ export function requestHandler(
       const scope = getCurrentScope();
       scope.setSDKProcessingMetadata({
         request: req,
-        // TODO (v8): Stop passing this
-        requestDataOptionsFromExpressHandler: requestDataOptions,
       });
 
       const client = getClient<NodeClient>();
@@ -372,7 +336,6 @@ export function trpcMiddleware(options: SentryTrpcMiddlewareOptions = {}) {
     }
 
     if (isThenable(maybePromiseResult)) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
       Promise.resolve(maybePromiseResult).then(
         nextResult => {
           captureIfError(nextResult as any);
@@ -389,9 +352,3 @@ export function trpcMiddleware(options: SentryTrpcMiddlewareOptions = {}) {
     return maybePromiseResult;
   };
 }
-
-// TODO (v8 / #5257): Remove this
-// eslint-disable-next-line deprecation/deprecation
-export type { ParseRequestOptions, ExpressRequest } from './requestDataDeprecated';
-// eslint-disable-next-line deprecation/deprecation
-export { parseRequest, extractRequestData } from './requestDataDeprecated';
