+++ packages/node/src/sdk.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import {
   endSession,
   functionToStringIntegration,
