+++ packages/serverless/src/gcpfunction/http.ts
@@ -5,7 +5,6 @@ import {
   handleCallbackErrors,
   setHttpStatus,
 } from '@sentry/core';
-import type { AddRequestDataToEventOptions } from '@sentry/node';
 import { continueTrace, startSpanManual } from '@sentry/node';
 import { getCurrentScope } from '@sentry/node';
 import { captureException, flush } from '@sentry/node';
@@ -15,24 +14,6 @@ import { DEBUG_BUILD } from '../debug-build';
 import { domainify, markEventUnhandled, proxyFunction } from './../utils';
 import type { HttpFunction, WrapperOptions } from './general';
 
-// TODO (v8 / #5257): Remove this whole old/new business and just use the new stuff
-type ParseRequestOptions = AddRequestDataToEventOptions['include'] & {
-  serverName?: boolean;
-  version?: boolean;
-};
-
-interface OldHttpFunctionWrapperOptions extends WrapperOptions {
-  /**
-   * @deprecated Use `addRequestDataToEventOptions` instead.
-   */
-  parseRequestOptions: ParseRequestOptions;
-}
-interface NewHttpFunctionWrapperOptions extends WrapperOptions {
-  addRequestDataToEventOptions: AddRequestDataToEventOptions;
-}
-
-export type HttpFunctionWrapperOptions = OldHttpFunctionWrapperOptions | NewHttpFunctionWrapperOptions;
-
 /**
  * Wraps an HTTP function handler adding it error capture and tracing capabilities.
  *
@@ -40,10 +21,7 @@ export type HttpFunctionWrapperOptions = OldHttpFunctionWrapperOptions | NewHttp
  * @param options Options
  * @returns HTTP handler
  */
-export function wrapHttpFunction(
-  fn: HttpFunction,
-  wrapOptions: Partial<HttpFunctionWrapperOptions> = {},
-): HttpFunction {
+export function wrapHttpFunction(fn: HttpFunction, wrapOptions: Partial<WrapperOptions> = {}): HttpFunction {
   const wrap = (f: HttpFunction): HttpFunction => domainify(_wrapHttpFunction(f, wrapOptions));
 
   let overrides: Record<PropertyKey, unknown> | undefined;
@@ -59,17 +37,7 @@ export function wrapHttpFunction(
 }
 
 /** */
-function _wrapHttpFunction(fn: HttpFunction, wrapOptions: Partial<HttpFunctionWrapperOptions> = {}): HttpFunction {
-  // TODO (v8 / #5257): Switch to using `addRequestDataToEventOptions`
-  // eslint-disable-next-line deprecation/deprecation
-  const { parseRequestOptions } = wrapOptions as OldHttpFunctionWrapperOptions;
-
-  const options: HttpFunctionWrapperOptions = {
-    flushTimeout: 2000,
-    // TODO (v8 / xxx): Remove this line, since `addRequestDataToEventOptions` will be included in the spread of `wrapOptions`
-    addRequestDataToEventOptions: parseRequestOptions ? { include: parseRequestOptions } : {},
-    ...wrapOptions,
-  };
+function _wrapHttpFunction(fn: HttpFunction, options: Partial<WrapperOptions> = { flushTimeout: 2000 }): HttpFunction {
   return (req, res) => {
     const reqMethod = (req.method || '').toUpperCase();
     const reqUrl = stripUrlQueryAndFragment(req.originalUrl || req.url || '');
@@ -90,7 +58,6 @@ function _wrapHttpFunction(fn: HttpFunction, wrapOptions: Partial<HttpFunctionWr
         span => {
           getCurrentScope().setSDKProcessingMetadata({
             request: req,
-            requestDataOptionsFromGCPWrapper: options.addRequestDataToEventOptions,
           });
 
           if (span instanceof Transaction) {
