+++ packages/utils/src/requestdata.ts
@@ -189,8 +189,6 @@ export function extractRequestData(
   req: PolymorphicRequest,
   options?: {
     include?: string[];
-    // TODO(v8): Remove this paramater
-    deps?: InjectedNodeDeps;
   },
 ): ExtractedNodeRequestData {
   const { include = DEFAULT_REQUEST_INCLUDES } = options || {};
@@ -258,7 +256,6 @@ export function extractRequestData(
         // query string:
         //   node: req.url (raw)
         //   express, koa, nextjs: req.query
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         requestData.query_string = extractQueryParams(req);
         break;
       }
@@ -309,8 +306,8 @@ export function addRequestDataToEvent(
 
   if (include.request) {
     const extractedRequestData = Array.isArray(include.request)
-      ? extractRequestData(req, { include: include.request, deps: options && options.deps })
-      : extractRequestData(req, { deps: options && options.deps });
+      ? extractRequestData(req, { include: include.request })
+      : extractRequestData(req);
 
     event.request = {
       ...event.request,
